Move static truthiness to new module
